Change font sizes to decimal
Using decimal to match recent changes to other size values.

Fixes #653

+++ src/ShapeCrawler/Fonts/IFontSize.cs
@@ -8,10 +8,10 @@ internal interface IFontSize
     /// <summary>
     ///     Font size in points.
     /// </summary>
-    int Size();
+    decimal Size();
     
     /// <summary>
     ///     Updates font size in points.
     /// </summary>
-    void Update(int points);
+    void Update(decimal points);
 }
